+++ crates/grafeo-adapters/src/query/gremlin/parser.rs
@@ -2,6 +2,7 @@
 //!
 //! Parses tokenized Gremlin queries into an AST.
 
+#[allow(clippy::wildcard_imports)]
 use super::ast::*;
 use super::lexer::{Lexer, Token, TokenKind};
 use grafeo_common::types::Value;