+++ crates/grafeo-adapters/src/query/sparql/parser.rs
@@ -2,6 +2,7 @@
 //!
 //! Implements a recursive descent parser for SPARQL 1.1 Query Language.
 
+#[allow(clippy::wildcard_imports)]
 use super::ast::*;
 use super::lexer::{Lexer, Token, TokenKind};
 use grafeo_common::utils::error::{Error, QueryError, QueryErrorKind, Result};